Rename files from tft to currency
